Create code_practice_33
check if a given string is valid or not. if it contains combinations of alphabets (in upper case or lower case) with/without spaces. 
If the string given as argument is "Hello Good Morning", the function will return True.
If the string given as argument is "Purabi@gmail.com", the function will return False.