Create code_practice_35
Write a python code to print the count of valid strings and invalid strings from a given list of strings.
A string is considered as valid if it contains combinations of alphabets (in upper case or lower case) with/without spaces.
Define a function to check if a given string is valid or not i.e if it contains combinations of alphabets (in upper case or lower case) with/without spaces. 
This function will take a string as input and return True if the string is valid, otherwise will return False.
Example :
If the string given as argument is "Hello Good Morning", the function will return True.
If the string given as argument is "Purabi@gmail.com", the function will return False.
Using this function in main section, build the logic to print the count of valid strings and invalid strings from a given list of strings.